requirements.txt
# Core Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
scipy>=1.11.0

# Options Data Sources
yahooquery>=2.3.0
yfinance>=0.2.20

# Options Calculations & Greeks
py_vollib>=1.0.1

# News & Sentiment Analysis
transformers>=4.30.0
torch>=2.0.0
beautifulsoup4>=4.12.0

# Technical Analysis
pandas-ta>=0.3.14b

# APIs
polygon-api-client>=1.12.0
aiohttp>=3.8.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0

# Web Scraping
lxml>=4.9.0

# Development
pytest>=7.4.0
black>=23.0.0
